feat: extract SessionStatus + SessionRunState to @opencode-ai/state
- Moved status.ts (97 lines) to state, zero @/-imports
- Moved run-state.ts (156 lines) to state, zero @/-imports
- Moved Runner utility (217 lines) to @opencode-ai/core/effect/runner
- All dependencies resolved to @opencode-ai/core equivalents
